pyproject.toml
[project]
name = "gecko"
version = "0.1.0"
description = "Members + Gecko opt-in announcement Discord bot with FastAPI server"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.4",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "aiohttp>=3.9",
    "python-dotenv>=1.0",
]

[tool.uv]
package = false
